refactor(agent-tools): extract link tools into agent-link-tools.ts, make agent-tools.ts a pure barrel
